refactor : Payment 메서드명 변경
