+++ src/test/java/lotto/PaymentTest.java
@@ -9,15 +9,13 @@ import org.junit.jupiter.api.Test;
 
 public class PaymentTest {
 
-    final static int LOTTO_PRICE = 1000;
-
     @DisplayName("투입한 금액이 1,000원 미만이면 예외 발생한다")
     @Test()
     void 최소_투입_금액_검증() {
         int value = 999;
 
         assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> new Payment(value, LOTTO_PRICE))
+            .isThrownBy(() -> new Payment(value))
             .withMessage("1,000원 이상 투입해 주세요");
     }
 
@@ -27,11 +25,11 @@ public class PaymentTest {
         int value1 = 12000;
         int value2 = 1000;
 
-        Payment payment1 = new Payment(value1, LOTTO_PRICE);
-        Payment payment2 = new Payment(value2, LOTTO_PRICE);
+        Payment payment1 = new Payment(value1);
+        Payment payment2 = new Payment(value2);
 
-        assertThat(payment1.dividedWith(LOTTO_PRICE)).isEqualTo(12);
-        assertThat(payment2.dividedWith(LOTTO_PRICE)).isEqualTo(1);
+        assertThat(payment1.calculateLottoCount()).isEqualTo(12);
+        assertThat(payment2.calculateLottoCount()).isEqualTo(1);
     }
 
 }
